refactor(gemini_client): extract Groq IO, postprocess, and context modules
Pulls the Groq client singleton, Whisper transcription, the agentic chat
+tool-call retry loop, and JSON extraction out of client.py into a private
_groq.py. The deterministic post-processing (singularize plural units,
normalize vague-qty, rewrite vague-qty clarification questions) moves to
postprocess.py with a single apply() entry point. _build_context becomes
context.assemble_context().

client.py shrinks from 257 → ~125 lines and is now pure orchestration:
transcribe → assemble_context → chat_with_tools → extract_json → apply →
validate. Behaviour and the system prompt are unchanged; this is the
extraction step before the router and per-mode handler split.

Why: the monolithic client.py blocks the router/handler refactor (issue
#36). Extracting the IO and post-processing first keeps the diff
reviewable and means the handler split touches only orchestration code.

Verified: all 16 mocked unit tests pass; live spot-check on
vague_splash_olive_oil / past_added_oil / noqty_i_want_milk shows
expected behaviour (vague pass, the two known baseline-locked
regressions still fail in the same way).
